Move folderId to DiscInput type
Having id of related object as separate mutation parameter works OK in case of one or two relations.
It gets messy with more related objects. In `SongInput` type we are going to have 3 relations: disc, artist and genre.
This change is a preparator stage for implementing song mutation, so that we have unified way for passing identifiers of related objects.