Remove separate DAL entities
This is an enabler for complex DB update scenarios.
For example, adding of song playback requires following steps:

1. Adding playback to Playbacks set.
2. Incrementing Song.PlaybacksCount;
3. Updating Song.LastPlaybackTime.
4. Returning id of created playback.

All these steps should be implemented in BL layer (i.e. SongsService).
We want to use Unit of Work pattern to save all changes in one transaction.
If BL and DAL operate with the different entity objects,
it becomes challenging retrieval of id for created object.