Restore original MP3 tag fields order
The tag fields are written to the MP3 file in the order they are set on the `Tag` object.
This does not matter for the functional aspect. However, the files will
differ in binary content. Checksum will also differ.
We restore original order used by legacy `MusicLibrary` to keep binary file content the same.